+++ src/gafaelfawr/handlers/index.py
@@ -1,11 +1,7 @@
 """Handlers for the app's root, ``/``."""
 
-from email.message import Message
-from importlib.metadata import metadata
-from typing import cast
-
 from fastapi import APIRouter
-from safir.metadata import Metadata, get_project_url
+from safir.metadata import Metadata, get_metadata
 
 from ..slack import SlackRouteErrorHandler
 
@@ -31,11 +27,6 @@ async def get_index() -> Metadata:
 
     By convention, this endpoint returns only the application's metadata.
     """
-    pkg_metadata = cast(Message, metadata("gafaelfawr"))
-    return Metadata(
-        name="gafaelfawr",
-        version=pkg_metadata["Version"],
-        description=pkg_metadata["Summary"],
-        repository_url=get_project_url(pkg_metadata, "Source"),
-        documentation_url=get_project_url(pkg_metadata, "Homepage"),
+    return get_metadata(
+        package_name="gafaelfawr", application_name="gafaelfawr"
     )