+++ src/gafaelfawr/handlers/login.py
@@ -72,7 +72,6 @@ class LoginError(Enum):
     summary="Authenticate browser",
     tags=["browser"],
 )
-@router.get("/oauth2/callback", include_in_schema=False, tags=["browser"])
 async def get_login(
     code: Optional[str] = Query(
         None,
@@ -267,14 +266,6 @@ async def handle_provider_return(
     # what they were doing.
     context.state.state = None
     context.state.return_url = None
-    context.logger.info(
-        "Successfully authenticated user %s (%s)",
-        user_info.username,
-        user_info.uid,
-        user=user_info.username,
-        token=token.key,
-        scopes=sorted(scopes),
-    )
     return RedirectResponse(return_url)